// src/api/openapi/model/offerFieldEnum.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const OfferFieldEnum = {
    Numeric: 'NUMERIC',
    String: 'STRING',
    Text: 'TEXT',
    Calculation: 'CALCULATION',
    Select: 'SELECT',
    Offertext: 'OFFERTEXT'
} as const;
export type OfferFieldEnum = typeof OfferFieldEnum[keyof typeof OfferFieldEnum];